requirements.txt
beautifulsoup4=4.8.0
requests=2.22.0
python-dateutil=2.8.0
pyyaml=5.1.2